Commit29: Corrected the issues NamedQuery and ID (s. the previous commit 883f63bd)

+++ springboot-multi/springboot-multi-persist/src/main/java/org/sonatype/mavenbook/repository/LocationRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface LocationRepository extends CrudRepository<Location, Long> {
+public interface LocationRepository extends CrudRepository<Location, Integer> {
     Optional<Location> findByWoeid(String woeid);
     Optional<Location> findByCity(String city);
 }
